Handling authentication exceptions gracefully and auto-refreshing tokens
Due to the nature of this application containing financial information, tokens will be short lived. Thus we do not really need a refresh token, and instead can rely on requests to automatically refresh tokens. There is a manual endpoint to refresh token before it expires. A mechanism/timer will be need in the frontend to manage this.